+++ src/DependencyInjection/SonataBlockExtension.php
@@ -158,6 +158,8 @@ final class SonataBlockExtension extends Extension
                 $contexts[$context][] = $service;
             }
         }
+
+        $container->setParameter('sonata_blocks.default_contexts', $defaults);
     }
 
     /**
@@ -234,6 +236,10 @@ final class SonataBlockExtension extends Extension
      */
     public function configureClassesToCompile(): void
     {
+        if (!\is_callable([$this, 'addClassesToCompile'])) {
+            return;
+        }
+
         $this->addClassesToCompile([
             'Sonata\\BlockBundle\\Block\\BlockLoaderChain',
             'Sonata\\BlockBundle\\Block\\BlockLoaderInterface',